+++ ChangeLog.markdown
@@ -5,6 +5,7 @@ Imbo-2.x.x
 ----------
 __N/A__
 
+* #436: Added new MongoDB adapters that use the mongodb extension (Christer Edvartsen)
 * #434: Custom models for group(s) and access rule(s) (Christer Edvartsen)
 
 Imbo-2.0.0